build: test publish paths before cleanup

+++ build.ps1
@@ -97,9 +97,15 @@ Write-Output "Building new xml file for help"
 New-ExternalHelp -Path "$PSScriptRoot\docs" -OutputPath "$PSScriptRoot\en-US\" -Force
 
 # Package for Publish
-Get-ChildItem "$PSScriptRoot/publish" -Recurse -File | Remove-Item -Force -Recurse
-Get-ChildItem $PSScriptRoot/publish/$moduleName/ -Recurse -Directory | Remove-Item -Force -Recurse
-Get-ChildItem $PSScriptRoot/publish -Recurse -Directory | Remove-Item -Force -Recurse
+if((Test-Path "$PSScriptRoot/publish")) {
+    Get-ChildItem "$PSScriptRoot/publish" -Recurse -File | Remove-Item -Force -Recurse
+}
+if((Test-Path "$PSScriptRoot/publish/$moduleName/")) {
+    Get-ChildItem $PSScriptRoot/publish/$moduleName/ -Recurse -Directory | Remove-Item -Force -Recurse
+}
+if((Test-Path "$PSScriptRoot/publish")) {
+    Get-ChildItem $PSScriptRoot/publish -Recurse -Directory | Remove-Item -Force -Recurse
+}
 New-Item -Path $PSScriptRoot/publish/$moduleName -ItemType Directory -Force -ErrorAction SilentlyContinue | Out-Null
 Copy-Item -Path $PSScriptRoot/docs -Destination $PSScriptRoot/publish/$moduleName/ -Recurse
 Copy-Item -Path $PSScriptRoot/en-US -Destination $PSScriptRoot/publish/$moduleName/ -Recurse
